feat: :sparkles: disable redirect from / to /ktsstudio
