+++ src/App/router/router.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { createBrowserRouter } from 'react-router-dom';
 import { MainPage, RepoPage, ErrorPage } from 'App/pages';
+import PaginationDisplay from 'App/pages/MainPage/components/PaginationDisplay';
 import Root from './Root';
 
 export const router = createBrowserRouter([
@@ -10,11 +11,12 @@ export const router = createBrowserRouter([
     errorElement: <ErrorPage />,
     children: [
       {
-        index: true,
         errorElement: <ErrorPage />,
-        path: ':org',
+        path: '',
         element: <MainPage />,
+        children: [{ path: ':org', element: <PaginationDisplay /> }],
       },
+
       {
         path: ':org/:repo',
         errorElement: <ErrorPage />,
